code refactor for the navigation bar
