+++ src/components/navigation/HeaderProgress.scss
@@ -1,4 +1,4 @@
-@import '../themes/common-variables';
+@import '../../themes/common-variables';
 
 .progressOuter {
   height: 0.25rem;
