+++ src/components/navigation/MobileNaviBar.scss
@@ -1,4 +1,4 @@
-@import '../themes/variables';
+@import '../../themes/variables';
 
 .mobileNav {
   transition: background-color 0.5s ease-in-out;
