+++ src/components/navigation/SubNavBar.scss
@@ -1,4 +1,4 @@
-@import '../themes/variables.scss';
+@import '../../themes/variables.scss';
 $item-height: 32px;
 
 .outer {
